+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.2.0
+
+## Added
+
+- `Client.stream_packages_v1` method to asynchronously stream packages
+
+## Changed
+
+- Renamed `PackageListing` and `PackageVersion` to `PackageV1` and `PackageVersionV1` respectively
+- Renamed `PackageExperimental` and `PackageVersionExperimental` to `Package` and `PackageVersion` respectively
+
 ## 0.1.0
 
 - Initial release